99.999% sure this should fix the list issue
Alrighty the issue is that all the buttons did was turn the things into a <ul> (unordered list) and an <ol> (ordered list). By default a <ul> is dotted and an <ol> is numbered (user agent stylesheet which is a browsers default settings) however the buttons didn't stylize the buttons, so some other css styles overrode the default settings. It can easily be fixed by simply adding a <style>